requirements.txt
django~=3.2.7
django-debug-toolbar~=3.2.2
requests~=2.26.0
beautifulsoup4~=4.10.0
djangorestframework~=3.12.4
aiohttp~=3.7.4.post0
django-filter~=2.4.0